final de la clase 29-11-2023

+++ src/PreparacionExamen2Ejercicios/bucleWhile/Ejercicio1.java
@@ -5,7 +5,7 @@ public class Ejercicio1 {
 //        Calcular el factorial de 25
         //1. declaración de variables
         int num= 25;
-        int factorial= 1;
+        long factorial= 1;
         while (num> 0){
             factorial *= num;
             num--;
